# backend/openapi.yaml
openapi: 3.0.0
info:
  title: AgentSpark Backend API
  version: 1.2.0
  description: Secure API proxy, authentication, and data persistence for AgentSpark
servers:
  - url: http://localhost:5000/api/v1
    description: Local development server
paths:
  /auth/register:
    post:
      summary: Register a new local development user via email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                name:
                  type: string
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    type: object
  /auth/google:
    post:
      summary: Login via Google OAuth credential 
      responses:
        '200':
          description: OK
  /auth/me:
    get:
      summary: Get details of the current authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile details returned
  /projects:
    get:
      summary: Get all projects for logged-in user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Array of generated agent projects
    post:
      summary: Create a new agent project
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Saved project ID
  /projects/{id}:
    get:
      summary: Get detailed project view by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project metadata and LLM-generated JSON
  /generate:
    post:
      summary: Proxy request to generate agents via LLM
      security:
        - bearerAuth: []
        - hmacSignature: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                topic:
                  type: string
                level:
                  type: string
                modelId:
                  type: string
      responses:
        '200':
          description: Streaming text LLM output payload resolving to list of Agents

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    hmacSignature:
      type: apiKey
      in: header
      name: X-Signature